requirements.txt
dialogflow==0.5.1
geopy==2.2.0
google-api-core==1.4.1
python-dotenv==0.20.0
